Desenvolvimento da parte do cadastro para criar e calcular a quanbtidade com seu preço, utiliznado javascript

+++ SistemaProduto.Dominio/Dominio/Produto.cs
@@ -13,7 +13,7 @@ namespace SistemaProduto.Dominio.Dominio
 
         public string? NomeProduto { get; set; }
 
-        public double Preco { get; set; }
+        public decimal Preco { get; set; }
 
         public int Quantidade { get; set; }
